+++ R/p_attribute.R
@@ -284,6 +284,16 @@ p_attribute_encode <- function(
       )
     }
   } else if (method == "R"){
+    
+    # Check whether corpus has been loaded and delete corpus if necessary
+    # cwb_makeall will crash if corpus is loaded
+    
+    corpus_size <- RcppCWB::cl_attribute_size(
+      corpus = corpus, attribute = "word", attribute_type = "p",
+      registry = registry_dir
+    )
+    if (corpus_size > 0L) cl_delete_corpus(corpus = corpus, registry = registry_dir)
+
     cwb_makeall(corpus = corpus, p_attribute = p_attribute, registry = registry_dir)
     if (compress){
       cwb_huffcode(corpus = corpus, p_attribute = p_attribute, registry = registry_dir)
@@ -302,8 +312,6 @@ p_attribute_encode <- function(
       }
     }
   }
-  
-  
 }